23 de julho

+++ resources/views/listagem_categoria.blade.php
@@ -9,12 +9,13 @@
   <body>
     <div class="container">
         <h1>Listagem de Categorias</h1>
-        <a href="novo">Nova Categoria</a>
+        <a class="btn btn-primary" href="novo">Nova Categoria</a>
         <table class="table table-bordered table-striped">
             <thead>
                 <tr>
                 <th scope="col">ID</th>
                 <th scope="col">Descrição</th>
+                <th scope="col">Editar</th>
                 <th scope="col">Excluir</th>
                 </tr>
             </thead>
@@ -23,7 +24,8 @@
                     <tr>
                     <th scope="row">{{ $categoria->id }}</th>
                     <td>{{$categoria->descricao}}</td>
-                    <td><a href="excluir/{{$categoria->id}}">-</a></td>
+                    <td><a class="btn btn-primary" href="editar/{{$categoria->id}}">+</a></td>
+                    <td><a class="btn btn-danger" onclick="return confirm('Confirme a remoção do registro?')" href="excluir/{{$categoria->id}}">-</a></td>
                     </tr>
                 @endforeach
             </tbody>
